ver.0.47.a: definition of real_pw has been changed into float from double.

+++ src/celeste/core/CelesteObject.h
@@ -11,8 +11,8 @@
 typedef double real;
 
 // type of real values for pairwise energy calculation in GPU
-// typedef float real_pw;
-typedef double real_pw;
+typedef float real_pw;
+//typedef double real_pw;
 
 // type of real values for summation of force,energy
 typedef double real_fc;
